next

+++ database/migrations/2024_04_04_170026_create_beban.php
@@ -14,7 +14,7 @@ class CreateBeban extends Migration
     public function up()
     {
         Schema::create('beban', function (Blueprint $table) {
-            $table->increments('idbeban');
+            $table->bigInteger('idbeban');
             $table->string('kode');
             $table->string('keterangan');
             $table->integer('biaya');
